Move `get_benchmarks` from the `collector` binary to the `benchmark` module
